+++ globalConfig.json
@@ -1,1 +1,1 @@
-{"mongoUri":"mongodb://127.0.0.1:43399/aca4debd-2318-4fed-b8cf-d441e58d6626?"}
+{"mongoUri":"mongodb://127.0.0.1:36247/e9ca1e47-b18c-46f5-a904-5b0668bd61d1?"}
