--- src/data/protocols/add-account-repository.ts
@@ -1,6 +1,0 @@
-import { AccountModel } from '../../domain/models/account'
-import { AddAccountModel } from '../../domain/useCases/add-account'
-
-export interface AddAccountRepository {
-  add: (accountData: AddAccountModel) => Promise<AccountModel>
-}
