+++ src/data/protocols/db/add-account-repository.ts
@@ -1,0 +1,6 @@
+import { AccountModel } from '../../../domain/models/account'
+import { AddAccountModel } from '../../../domain/useCases/add-account'
+
+export interface AddAccountRepository {
+  add: (accountData: AddAccountModel) => Promise<AccountModel>
+}
