+++ src/main/decorators/log.ts
@@ -1,5 +1,5 @@
 import { Controller, HttpRequest, HttpResponse } from '../../presentation/protocols'
-import { LogErrorRepository } from '../../data/protocols/log-error-repository'
+import { LogErrorRepository } from '../../data/protocols/db/log-error-repository'
 
 export class LogControllerDecorator implements Controller {
   private readonly controller: Controller
